requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
coverage==5.5
crispy-bootstrap5==0.4
Django==3.2.4
django-bootstrap-v5==1.0.4
django-crispy-forms==1.12.0
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
